backend/requirements.txt
fastapi
uvicorn
python-dotenv
elasticsearch
google-genai
assemblyai
requests
langfuse
# --- Add these new lines ---
docker
pyyaml